Implements basic health and death system
Adds a basic health system for entities using ECS.

Introduces a `HealthData` component and a `CheckDeathSystem` that destroys entities when their health reaches zero.

Also introduces `HealthAuthoring` to add the component to prefabs.

Finally it updates the `ShootingSystem` to use the new health component and also adds a `ResetTargetSystem` to reset the target if the target entity has been destroyed.